requirements.txt
altair==5.5.0
attrs==26.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.19.0
numpy==1.26.4
packaging==24.2
pandas==2.3.3
pillow==10.4.0
protobuf==4.25.9
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.20.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.31.0
rich==13.9.4
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
streamlit==1.33.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==4.0.0
